chore: move `normalizeSelector` out of engine
